refactor plugin location and name
