+++ plugins/catalog-backend-module-tyk/package.json
@@ -1,0 +1,37 @@
+{
+  "name": "@internal/plugin-catalog-backend-module-tyk",
+  "description": "The tyk backend module for the catalog plugin.",
+  "version": "0.1.0",
+  "main": "src/index.ts",
+  "types": "src/index.ts",
+  "license": "Apache-2.0",
+  "private": true,
+  "publishConfig": {
+    "access": "public",
+    "main": "dist/index.cjs.js",
+    "types": "dist/index.d.ts"
+  },
+  "backstage": {
+    "role": "backend-plugin-module"
+  },
+  "scripts": {
+    "start": "backstage-cli package start",
+    "build": "backstage-cli package build",
+    "lint": "backstage-cli package lint",
+    "test": "backstage-cli package test",
+    "clean": "backstage-cli package clean",
+    "prepack": "backstage-cli package prepack",
+    "postpack": "backstage-cli package postpack"
+  },
+  "dependencies": {
+    "@backstage/backend-common": "^0.19.8",
+    "@backstage/backend-plugin-api": "^0.6.6"
+  },
+  "devDependencies": {
+    "@backstage/backend-test-utils": "^0.2.7",
+    "@backstage/cli": "^0.23.1"
+  },
+  "files": [
+    "dist"
+  ]
+}
