+++ plugins/catalog-backend-module-tyk/src/clients/index.ts
@@ -1,0 +1,1 @@
+export * from './tyk-dashboard-client'
