+++ plugins/catalog-backend-module-tyk/src/index.ts
@@ -1,0 +1,8 @@
+/***/
+/**
+ * The tyk backend module for the catalog plugin.
+ *
+ * @packageDocumentation
+ */
+
+export { catalogModuleTyk } from './module';
