+++ plugins/catalog-backend-module-tyk/src/module.ts
@@ -1,0 +1,14 @@
+import { coreServices, createBackendModule } from '@backstage/backend-plugin-api';
+
+export const catalogModuleTyk = createBackendModule({
+  pluginId: 'catalog',
+  moduleId: 'tyk',
+  register(reg) {
+    reg.registerInit({
+      deps: { logger: coreServices.logger },
+      async init({ logger }) {
+        logger.info('Hello World!')
+      },
+    });
+  },
+});
